libs/soorma-nats/pyproject.toml
[project]
name = "soorma-nats"
version = "0.1.0"
description = "Shared NATS client for Soorma infrastructure services"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.6.0",
    "soorma-common",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soorma_nats"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
